+++ flex-objects.php
@@ -13,12 +13,18 @@ use Grav\Events\PermissionsRegisterEvent;
 use Grav\Events\PluginsLoadedEvent;
 use Grav\Framework\Acl\PermissionsReader;
 use Grav\Framework\Flex\FlexDirectory;
+use Grav\Framework\Flex\FlexForm;
+use Grav\Framework\Flex\Interfaces\FlexAuthorizeInterface;
 use Grav\Framework\Flex\Interfaces\FlexInterface;
+use Grav\Framework\Form\Interfaces\FormInterface;
+use Grav\Framework\Route\Route;
 use Grav\Plugin\Admin\Admin;
+use Grav\Plugin\FlexObjects\Controllers\ObjectController;
 use Grav\Plugin\FlexObjects\FlexFormFactory;
 use Grav\Plugin\Form\Forms;
 use Grav\Plugin\FlexObjects\Admin\AdminController;
 use Grav\Plugin\FlexObjects\Flex;
+use Psr\Http\Message\ServerRequestInterface;
 use RocketTheme\Toolbox\Event\Event;
 use function is_callable;
 
@@ -83,6 +89,30 @@ class FlexObjectsPlugin extends Plugin
             'onFormRegisterTypes' => [
                 ['onFormRegisterTypes', 0]
             ]
+        ];
+    }
+
+    /**
+     * Get list of form field types specified in this plugin. Only special types needs to be listed.
+     *
+     * @return array
+     */
+    public function getFormFieldTypes()
+    {
+        return [
+            'list' => [
+                'array' => true
+            ],
+            'pagemedia' => [
+                'array' => true,
+                'media_field' => true,
+                'validate' => [
+                    'type' => 'ignore'
+                ]
+            ],
+            'filepicker' => [
+                'media_picker_field' => true
+            ],
         ];
     }
 
@@ -172,6 +202,15 @@ class FlexObjectsPlugin extends Plugin
                 'onTwigTemplatePaths' => [
                     ['onTwigTemplatePaths', 0]
                 ],
+                'onPageInitialized' => [
+                    ['authorizePage', 10000]
+                ],
+                'onBeforeFlexFormInitialize' => [
+                    ['onBeforeFlexFormInitialize', -10]
+                ],
+                'onPageTask' => [
+                    ['onPageTask', -10]
+                ],
             ]);
         }
     }
@@ -206,6 +245,141 @@ class FlexObjectsPlugin extends Plugin
         $debugger = Grav::instance()['debugger'];
         $names = implode(', ', array_keys($flex->getDirectories()));
         $debugger->addMessage(sprintf('Registered flex types: %s', $names), 'debug');
+    }
+
+    /**
+     * @param Event $event
+     */
+    public function onBeforeFlexFormInitialize(Event $event): void
+    {
+        /** @var array $form */
+        $form = $event['form'];
+        $edit = $form['actions']['edit'] ?? false;
+        if (!isset($form['flex']['key']) && $edit === true) {
+            /** @var Route $route */
+            $route = $this->grav['route'];
+            $id = $route->getGravParam('id');
+            if (null !== $id) {
+                $form['flex']['key'] = $id;
+                $event['form'] = $form;
+            }
+        }
+    }
+
+    /**
+     * [onPageInitialized:10000] Authorize Flex Objects Page
+     */
+    public function authorizePage(Event $event): void
+    {
+        /** @var PageInterface|null $page */
+        $page = $event['page'];
+        if (null === $page) {
+            return;
+        }
+
+        $header = $page->header();
+        $forms = $page->forms();
+        $form = reset($forms);
+        if (($form['type'] ?? null) !== 'flex') {
+            $form = null;
+        }
+
+        // Make sure the page contains flex.
+        $config = $header->flex ?? [];
+        if (!$config && !$form) {
+            return;
+        }
+
+        /** @var Route $route */
+        $route = $this->grav['route'];
+
+        $type = $form['flex']['type'] ?? $config['directory'] ?? $route->getGravParam('directory') ?? null;
+        $key = $form['flex']['key'] ?? $config['id'] ?? $route->getGravParam('id') ?? '';
+        if (\is_string($type)) {
+            /** @var Flex $flex */
+            $flex = $this->grav['flex_objects'];
+            $directory = $flex->getDirectory($type);
+        } else {
+            $directory = null;
+        }
+
+        if (!$directory) {
+            return;
+        }
+
+        $create = (bool)($form['actions']['create'] ?? false);
+        $edit = (bool)($form['actions']['edit'] ?? false);
+
+        $scope = $config['access']['scope'] ?? null;
+
+        $object = $key !== '' ? $directory->getObject($key) : null;
+        $hasAccess = null;
+
+        $action = $config['access']['action'] ?? null;
+        if (null === $action) {
+            if (!$form) {
+                $action = $key !== '' ? 'read' : 'list';
+                if (null === $scope) {
+                    $hasAccess = true;
+                }
+            } elseif ($object) {
+                if ($edit) {
+                    $scope = $scope ?? 'admin';
+                    $action = 'update';
+                } else {
+                    $hasAccess = false;
+                }
+            } elseif ($create) {
+                $object = $directory->createObject([], $key);
+                $scope = $scope ?? 'admin';
+                $action = 'create';
+            } else {
+                $hasAccess = false;
+            }
+        }
+
+        if ($action && $hasAccess === null) {
+            if ($object instanceof FlexAuthorizeInterface) {
+                $hasAccess = $object->isAuthorized($action, $scope);
+            } else {
+                $hasAccess = $directory->isAuthorized($action, $scope);
+            }
+        }
+
+        if ($hasAccess) {
+            $page->modifyHeader('access', []);
+        } else {
+            $page->modifyHeader('access', ['admin.flex.no_access' => true]);
+        }
+    }
+
+    /**
+     * @param Event $event
+     */
+    public function onPageTask(Event $event): void
+    {
+        /** @var FormInterface|null $form */
+        $form = $event['form'] ?? null;
+        if (!$form instanceof FlexForm) {
+            return;
+        }
+
+        $object = $form->getObject();
+
+        /** @var ServerRequestInterface $request */
+        $request = $event['request'];
+        $request = $request
+            ->withAttribute('type', $object->getFlexType())
+            ->withAttribute('key', $object->getKey())
+            ->withAttribute('object', $object)
+            ->withAttribute('form', $form);
+
+        $controller = new ObjectController();
+
+        $response = $controller->handle($request);
+        if ($response->getStatusCode() !== 418) {
+            $this->grav->close($response);
+        }
     }
 
     /**