+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v1.1.0
+## 03/31/2021
+
+1. [](#new)
+   * Require **Grav 1.7.19** and **Form 5.1.0**
+   * Added basic frontend editing support
+   * Added `onBeforeFlexFormInitialize` event to help to initialize the frontend form
+1. [](#bugfix)
+   * Fixed error in admin when field validation fails
+
 # v1.0.16
 ## 07/19/2021